# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Fluxa API
  version: 1.0.0
  description: 'Fluxa is a multi-currency digital asset platform built on Stellar. It provides

    wallets, instant transfers, FX conversions, scheduled recurring payments,

    batch disbursements, fiat on/off-ramps, and a developer-friendly webhook system.


    All endpoints under `/v1` require a Bearer API key in the `Authorization` header.


    ## API key scopes


    An API key can be limited to specific resources and actions. Scopes have the

    form `<resource>:<action>`, where action is `read` or `write`:


    | Resource | Scopes | Covers |

    |---|---|---|

    | wallets | `wallets:read`, `wallets:write` | `/v1/wallets` (create, trustlines, contract wallet changes) |

    | transfers | `transfers:read`, `transfers:write` | `/v1/transfers`, `/v1/transactions`, `/v1/schedules`, `/v1/refunds` |

    | batches | `batches:read`, `batches:write` | `/v1/transfers/batch` |

    | webhooks | `webhooks:read`, `webhooks:write` | `/v1/webhooks` |

    | reports | `reports:read`, `reports:write` | `/v1/usage` (`reports:write` is reserved for report generation) |


    Other scopes: `keys:*`, `audit:read`, `fiat:*`, `fx:*`, `fees:read`,

    `beneficiaries:*`, `compliance:*`. `<resource>:*` grants both actions on one

    resource and `*` grants everything.


    On scoped resource groups, `GET`/`HEAD` need the `read` scope and every other

...

...

...

    path only; never the key itself).


...

...

...

...

...

...

    '
  contact:
...
  license:
...
servers:
...
security:
- BearerAuth: []
paths:
...
  /v1/auth/register:
    post:
      operationId: registerTenant
      summary: Register a new tenant
      description: Creates a new tenant account and returns the initial API key.
      tags:
      - Auth
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              name: Acme Corp
              email: ops@acme.com
      responses:
        '201':
          description: Tenant registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
              example:
                tenant_id: t_8f3a2b1c
                api_key: sk_live_abc123def456
                name: Acme Corp
                email: ops@acme.com
                created_at: '2026-01-15T10:00:00Z'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/auth/login:
    post:
      operationId: loginTenant
      summary: Login and issue a new API key
      description: Authenticates an existing tenant and returns a new API key.
      tags:
      - Auth
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: ops@acme.com
              password: s3cret
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
              example:
                tenant_id: t_8f3a2b1c
                api_key: sk_live_abc123def456
                name: Acme Corp
                email: ops@acme.com
                created_at: '2026-01-15T10:00:00Z'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /v1/withdrawals:
    post:
      operationId: initiateWithdrawal
      summary: Initiate a withdrawal
      description: Converts USDC from the wallet and sends fiat to a bank account, protected by idempotency.
      tags:
      - Wallets
      parameters:
      - name: X-Idempotency-Key
        in: header
        required: false
        schema:
          type: string
        description: Optional idempotency key to prevent duplicate withdrawals on network retries
      - name: Idempotency-Key
        in: header
        required: false
        schema:
          type: string
          format: uuid
        description: Optional idempotency key to prevent duplicate withdrawals on network retries
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - wallet_id
              - amount
              - currency
              - account_bank
              - account_number
              properties:
                wallet_id:
                  type: string
                  format: uuid
                amount:
                  type: string
                currency:
                  type: string
                account_bank:
                  type: string
                account_number:
                  type: string
      responses:
        '200':
          description: Withdrawal initiated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WithdrawalResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token
      description: 'Pass a tenant API key (`sk_live_...` or `sk_test_...`) or JWT bearer token.

        In Swagger UI, enter your token in the Authorize dialog (the ''Bearer '' prefix is added automatically).

        '
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: 'Pass an API key as `Bearer <key>` in the `Authorization` header.

        '
    AdminAuth:
      type: http
      scheme: bearer
      bearerFormat: admin token
      description: Administrative bearer token.
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: BAD_REQUEST
              message: invalid request body
              status: 400
              request_id: req-example
    Unauthorized:
      description: Unauthorized — missing or invalid API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: UNAUTHORIZED
              message: missing or invalid authorization header
              status: 401
              request_id: req-example
...
    Conflict:
      description: Conflict — a request with this Idempotency-Key is still in flight
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: REQUEST_IN_PROGRESS
              message: a request with this Idempotency-Key is still in progress
              status: 409
              request_id: req-example
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: INTERNAL_ERROR
              message: an unexpected error occurred
              status: 500
              request_id: req-example
  schemas:
...
    ErrorResponse:
      type: object
      required:
      - error
      properties:
        error:
          type: object
          required:
          - code
          - message
          - status
          - request_id
          properties:
            code:
              type: string
              enum:
              - BAD_REQUEST
              - ANCHOR_DEPOSIT_FAILED
              - ANCHOR_REGISTRATION_FAILED
              - ANCHOR_TRANSACTION_NOT_FOUND
              - ANCHOR_WITHDRAWAL_FAILED
              - API_KEY_NOT_FOUND
              - API_KEY_REVOKED
              - API_KEY_EXPIRED
              - BENEFICIARY_EXISTS
              - BENEFICIARY_NOT_ACTIVE
              - BENEFICIARY_NOT_ALLOWED
              - CONFIG_NOT_FOUND
              - CONFLICT
              - COOLING_OFF
              - CROSS_MODE_KEY_OPERATION
              - ENVIRONMENT_REQUIRED
              - IDEMPOTENCY_KEY_REQUIRED
              - IDEMPOTENCY_RECORD_PERSISTENCE_FAILED
              - IDEMPOTENCY_RESPONSE_TOO_LARGE
              - INSUFFICIENT_SCOPE
              - INVALID_ENVIRONMENT_MODE
              - INVALID_PREDICATE
              - INVALID_REQUEST
              - LAST_ORG_OWNER
              - METHOD_NOT_ALLOWED
              - MISSING_TRUSTLINE
              - NOT_IMPLEMENTED
              - PAYMENT_LINK_UNAVAILABLE
              - PREDICATE_NOT_SATISFIABLE
              - RATE_LIMITED
              - REFUND_CONFLICT
              - TEST_MODE_REQUIRED
              - TRANSFER_ID_INVALID
              - TRANSFER_ID_REQUIRED
              - TRANSFER_NOT_CANCELLABLE
              - UNSUPPORTED_FIAT_CURRENCY
              - WALLET_ID_REQUIRED
              - NOT_FOUND
              - UNAUTHORIZED
              - FORBIDDEN
              - INSUFFICIENT_SCOPE
              - INTERNAL_ERROR
              - INSUFFICIENT_SWEEPABLE_BALANCE
              - QUOTE_EXPIRED
              - QUOTE_ALREADY_USED
              - REVIEW_ALREADY_DECIDED
              - CLAIMABLE_BALANCE_NOT_PENDING
              - REQUEST_IN_PROGRESS
              description: Machine-readable error code
            message:
              type: string
              description: Human-readable error message
            status:
              type: integer
              description: HTTP status code; matches the response status
              minimum: 400
              maximum: 599
            request_id:
              type: string
              description: Effective X-Request-ID response header value
...
    RegisterRequest:
      type: object
      required:
      - name
      - email
      - password
      properties:
        name:
          type: string
          description: Tenant / organization name
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 8
    LoginRequest:
      type: object
      required:
      - email
      - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    AuthResponse:
      type: object
      properties:
        tenant_id:
          type: string
        api_key:
          type: string
          description: Raw API key — returned once only
        name:
          type: string
        email:
          type: string
          format: email
        created_at:
          type: string
          format: date-time
...
    WithdrawalResponse:
      type: object
      properties:
        reference:
          type: string
          description: Unique withdrawal reference (WIT-XXXXXXXX)
        status:
          type: string
          enum:
          - pending
          - completed
          - failed
...